// projet-comparatif-voitures-fullstack/src/app/models/car.ts
export class Car {
    _id?: number | string;
    brand: string;
    mileage: number;
    numberOfDoors: number;
    numberOfSeats: number;
    transmission: string;
    color: string;
    electric: boolean;
    category: string;
    price: number;
    rating: number;
    quantity: number;
    city: string;
}